Split parse logic for format task into new module
